+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - docs for how to contribute a new dataset (#46)
+- new `MDAnalysis.__authors__` attribute with the list of AUTHORS
 
 ### Changes
 - update online docs theme (#43)